+++ .editorconfig
@@ -1,0 +1,30 @@
+# EditorConfig helps developers define and maintain consistent
+# coding styles between different editors and IDEs
+# http://editorconfig.org
+
+root = true
+
+[*]
+indent_style = tab
+indent_size = 4
+
+end_of_line = lf
+charset = utf-8
+trim_trailing_whitespace = true
+insert_final_newline = true
+
+
+[*.{json}]
+indent_style = tab
+indent_size = 4
+
+
+[package.json]
+# package.json adhere to --save-dev manipulation styles
+indent_style = space
+indent_size = 2
+
+
+[*.{md,yml}]
+indent_style = space
+indent_size = 2
