+++ src/components/Helper/UploadImg/style.module.less
@@ -1,0 +1,37 @@
+.readonly_input_field{
+  background-color: @bg-input-field !important;
+  padding: 7px;
+  width: 100%;
+  cursor: pointer;
+}
+
+.readonly_input_field_upload{
+  background-color: @bg-input-field !important;
+  padding: 7px;
+  width: 100%;
+}
+
+.upload > div{
+  width: 100%;
+  cursor: pointer;
+}
+
+.btn_delete{
+  padding: 18px 0 18px 0;
+  color: rgba(0, 0, 0, 0.7);
+}
+
+.btn_delete:hover {
+  border-color: #ff7875 !important;
+  color: #ff4d4f !important;
+}
+
+.btn_view{
+  padding: 18px 0 18px 0;
+  color: rgba(0, 0, 0, 0.7);
+}
+
+.btn_view:hover {
+  border-color: #40a9ff !important;
+  color: #1890ff !important;
+}
