+++ src/components/ObjectPembiayaan/style.module.less
@@ -9,51 +9,3 @@
 .input_field_ec{
   padding: 7px;
 }
-
-.readonly_input_field{
-  background-color: @bg-input-field !important;
-  padding: 7px;
-  width: 100%;
-  cursor: pointer;
-}
-
-.readonly_input_field_upload{
-  background-color: @bg-input-field !important;
-  padding: 7px;
-  width: 100%;
-}
-
-.upload > div{
-  width: 100%;
-  cursor: pointer;
-}
-
-.btn_download{
-  padding: 18px 0 18px 0;
-  color: rgba(0, 0, 0, 0.7);
-}
-
-.btn_download:hover {
-  border-color: #73d13d !important;
-  color: #52c41a !important;
-}
-
-.btn_delete{
-  padding: 18px 0 18px 0;
-  color: rgba(0, 0, 0, 0.7);
-}
-
-.btn_delete:hover {
-  border-color: #ff7875 !important;
-  color: #ff4d4f !important;
-}
-
-.btn_view{
-  padding: 18px 0 18px 0;
-  color: rgba(0, 0, 0, 0.7);
-}
-
-.btn_view:hover {
-  border-color: #40a9ff !important;
-  color: #1890ff !important;
-}
